requirements.txt
# Core ML and Deep Learning (CPU-only)
torch==2.4.1
torchvision==0.19.1

# Image Processing
Pillow==10.4.0
opencv-python-headless==4.12.0.88
numpy==1.24.3

# Web Interface
streamlit==1.28.2

# Data Science and Visualization
matplotlib==3.7.5
seaborn==0.12.2
pandas==2.0.3
scikit-learn==1.3.2

# Configuration and Utilities
PyYAML==6.0.2
tqdm==4.67.1

# Federated Learning (optional)
flwr==1.11.1

# Additional dependencies for medical imaging
scikit-image==0.21.0
